+++ src/main/java/wintersteve25/tau/Tau.java
@@ -1,6 +1,7 @@
 package wintersteve25.tau;
 
 import net.minecraftforge.fml.common.Mod;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;