+++ src/main/java/wintersteve25/tau/renderer/ScreenUIRenderer.java
@@ -4,6 +4,7 @@ import com.mojang.blaze3d.matrix.MatrixStack;
 import net.minecraft.client.gui.IRenderable;
 import net.minecraft.client.gui.screen.Screen;
 import net.minecraft.util.text.StringTextComponent;
+import net.minecraftforge.fml.client.gui.GuiUtils;
 import wintersteve25.tau.components.base.DynamicUIComponent;
 import wintersteve25.tau.build.UIBuilder;
 import wintersteve25.tau.components.base.UIComponent;