+++ src/main/java/wintersteve25/tau/tests/ClientEvents.java
@@ -15,7 +15,7 @@ public class ClientEvents {
     @SubscribeEvent
     public static void onKeyDown(InputEvent.KeyInputEvent evet) {
         if (evet.getKey() == GLFW.GLFW_KEY_G) {
-            Minecraft.getInstance().setScreen(new ScreenUIRenderer(new TestScroll()));
+            Minecraft.getInstance().setScreen(new ScreenUIRenderer(new TestTooltip()));
         }
     }