+++ src/main/java/wintersteve25/tau/tests/TestTooltip.java
@@ -1,0 +1,17 @@
+package wintersteve25.tau.tests;
+
+import net.minecraft.util.text.StringTextComponent;
+import wintersteve25.tau.components.Center;
+import wintersteve25.tau.components.Text;
+import wintersteve25.tau.components.Tooltip;
+import wintersteve25.tau.components.base.UIComponent;
+import wintersteve25.tau.layout.Layout;
+
+public class TestTooltip implements UIComponent {
+    @Override
+    public UIComponent build(Layout layout) {
+        return new Center(new Tooltip.Builder()
+                .withText(new StringTextComponent("Helloooooo"))
+                .build(new Text.Builder("Hello!")));
+    }
+}